app/requirements.txt
aiogram
psycopg2-binary
python-dotenv
requests
openai
pybitrix